requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
antlr4-python3-runtime==4.9.3
asgiref==3.7.2
asteroid-filterbanks==0.4.0
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
azure-cognitiveservices-speech==1.36.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.4.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.8.2
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Django==5.0.2
django-multiupload==0.6.1
docopt==0.6.2
einops==0.7.0
executing==2.0.1
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
huggingface-hub==0.21.4
HyperPyYAML==1.2.2
idna==3.4
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
julius==0.2.7
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
kombu==5.3.7
lazy_loader==0.3
librosa==0.10.1
lightning==2.2.1
lightning-utilities==0.10.1
llvmlite==0.42.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
omegaconf==2.3.0
optuna==3.5.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.0
primePy==1.3
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.21
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1
pytorch-metric-learning==2.4.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.4
regex==2023.12.25
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.28
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.15.2
torch==2.2.1+cu121
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.2
torchmetrics==1.3.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
webrtcvad==2.0.10
yarl==1.9.4
